+++ contracts/compound/interfaces/IMaker.sol
@@ -4,31 +4,47 @@ pragma solidity ^0.8.13;
 /*** Maker Interfaces ***/
 
 interface PotLike {
-    function chi() external view returns (uint);
-    function dsr() external view returns (uint);
-    function rho() external view returns (uint);
-    function pie(address) external view returns (uint);
-    function drip() external returns (uint);
-    function join(uint) external;
-    function exit(uint) external;
+    function chi() external view returns (uint256);
+
+    function dsr() external view returns (uint256);
+
+    function rho() external view returns (uint256);
+
+    function pie(address) external view returns (uint256);
+
+    function drip() external returns (uint256);
+
+    function join(uint256) external;
+
+    function exit(uint256) external;
 }
 
 /*** Maker Interfaces for CDaiDelegate.sol ***/
 
 interface GemLike {
-    function approve(address, uint) external;
-    function balanceOf(address) external view returns (uint);
-    function transferFrom(address, address, uint) external returns (bool);
+    function approve(address, uint256) external;
+
+    function balanceOf(address) external view returns (uint256);
+
+    function transferFrom(
+        address,
+        address,
+        uint256
+    ) external returns (bool);
 }
 
 interface VatLike {
-    function dai(address) external view returns (uint);
+    function dai(address) external view returns (uint256);
+
     function hope(address) external;
 }
 
 interface DaiJoinLike {
     function vat() external returns (VatLike);
+
     function dai() external returns (GemLike);
-    function join(address, uint) external payable;
-    function exit(address, uint) external;
+
+    function join(address, uint256) external payable;
+
+    function exit(address, uint256) external;
 }
